+++ src/mgt-core/utils/SdkVersionMiddleware.ts
@@ -7,7 +7,7 @@
 
 import { Context, Middleware } from '@microsoft/microsoft-graph-client';
 import { getRequestHeader, setRequestHeader } from '@microsoft/microsoft-graph-client/lib/es/middleware/MiddlewareUtil';
-import { ComponentMiddlewareOptions } from '..';
+import { ComponentMiddlewareOptions } from './ComponentMiddlewareOptions';
 
 /**
  * Implements Middleware for the Graph sdk to inject
